requirements.txt
# requirements.txt python package requirements for the flaskapis

# Copyright (c) 2015 Rubens S. Gomes
# All Rights Reserved.

# ATTENTION: the user's environment locale should be set to us_EN.UTF8
# ATTENTION: the following list should also be defined in setup.py

arrow>=0.5.4
enum34>=1.0.4
nose>=1.3.4
requests>=2.6.0
sure>=1.2.9
validate_email>=1.2
Flask>=0.10.1
Flask-RESTful>=0.3.2
Pint>=0.6
python-daemon>=2.0.5
w1thermsensor>=0.2.1
paho-mqtt>=1.1
beautifulsoup4>=4.4.0
#Following is only needed on the raspberry pi
#RPi.GPIO>=0.5.11

# Fabric is needed on the remote client machine where builds are deployed from.
#Fabric>=1.10.1
